requirements.txt
dj-database-url==0.4.2
Django==1.11.6
django-cors-headers==2.1.0
djangorestframework==3.7.1
gunicorn==19.7.1
psycopg2==2.7.3.1
pytz==2017.2
